Time: 1 ms (98.80%), Space: 39.7 MB (84.59%) - LeetHub

+++ 69-sqrtx/69-sqrtx.java
@@ -1,7 +1,7 @@
 import java.lang.Math;
 class Solution {
     public int mySqrt(int x) {
-        int a=(int)Math.sqrt(x);
-        return a;
+        double a=(int)Math.sqrt(x);
+        return (int)a;
     }
 }
